requirements.txt
# Core dependencies
smolagents
cloudscraper
beautifulsoup4
transformers
torch
streamlit
python-dotenv
openai
google-generativeai
